reading and writing controller mapping reimplemented This fixes one of the regressions introduced with the input system rewrite. Controller input mappings are now stored in a subdirectory input.d and are read from files *.cfg in lexical order. Usually the files would be named starting with numbers such as 50-blabla.cfg. This scheme is also used when the program saves learned input configurations.
Unfortunately old input.cfg files can not imported into the new format
because the new system stores configs separate for individual controllers
identified by their system name. The old system only has mappings for the
controller numbers (i.e. 1st controller, 2nd controller,... based on the
order they appear on the system, usually the order they are plugged in).
The only thing that could be done with the old input.cfg is to use it as
default mapping for ALL unknown controllers. However that would be way
worse than the new built-in default mapping because it would permanently
break the ability to work out of the box with new controllers and dance
pads the user may buy later.

The one regression that still exists is that the GUI for changing the
mappings has not been re-written, yet, so atm you can only change the
mappings by editing input.d files by hand.